fix signed vs unsigned comparision warning

+++ place.cpp
@@ -182,7 +182,7 @@ void force_directed(std::vector<cell_t>& cells, rows_t& rows)
 
                                 int dist = abs(target_pos.y - y) + abs(target_pos.x - x);
 
-                                if ((dist == best_dist && target_pos.y == y) ||
+                                if ((dist == best_dist && target_pos.y == (signed int)y) ||
                                     (dist < best_dist)) {
                                     best_dist = dist;
                                     best_pos.x = x;
